feat: replace SoftReferencePooledObjectArray.kt with PlayerInfoRepository as it is too specialized anyway, and type-casting causes performance hits
